% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_msaeRB.R
\name{est_msaeRB}
\alias{est_msaeRB}
\title{EBLUPs Ratio Benchmarking based on a Multivariate Fay Herriot (Model 1)}
\usage{
est_msaeRB(
  formula,
  vardir,
  weight,
  samevar = FALSE,
  MAXITER = 100,
  PRECISION = 1e-04,
  data
)
}
\arguments{
\item{formula}{an object of class list of formula describe the fitted models}

\item{vardir}{matrix containing sampling variances of direct estimators. The order is: 'var1, cov12, ..., cov1r, var2, cov23, ..., cov2r, ..., cov(r-1)(r), var(r)'}

\item{weight}{matrix containing proportion of units in small areas. The order is: 'w1, w2, ..., w(r)'}

\item{samevar}{logical. If TRUE, the varians is same. Default is FALSE}

\item{MAXITER}{maximum number of iterations for Fisher-scoring. Default is 100}

\item{PRECISION}{coverage tolerance limit for the Fisher Scoring algorithm. Default value is 1E-4}

\item{data}{dataframe containing the variables named in formula, vardir, and weight}
}
\value{
dataframe
}
\description{
This function gives EBLUPs ratio benchmarking based on multivariate Fay-Herriot (Model 1)
}
