% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_saeRB.R
\name{est_saeRB}
\alias{est_saeRB}
\title{EBLUPs Ratio Benchmarking based on a Univariate Fay-Herriot (Model 1)}
\usage{
est_saeRB(
  formula,
  vardir,
  weight,
  samevar = FALSE,
  MAXITER = 100,
  PRECISION = 1e-04,
  data
)
}
\arguments{
\item{formula}{an object of class list of formula describe the fitted model}

\item{vardir}{vector containing sampling variances of direct estimators}

\item{weight}{vector containing proportion of units in small areas}

\item{samevar}{logical. If TRUE, the varians is same. Default is FALSE}

\item{MAXITER}{maximum number of iterations for Fisher-scoring. Default is 100}

\item{PRECISION}{coverage tolerance limit for the Fisher Scoring algorithm. Default value is 1E-4}

\item{data}{dataframe containing the variables named in formula, vardir, and weight}
}
\value{
dataframe
}
\description{
This function gives EBLUPs ratio benchmarking based on univariate Fay-Herriot (model 1)
}
